alle uints gefixd?
git-svn-id: https://psvn.ims.uni-hannover.de/svn/fucap/trunk@132 1233f08c-2788-c397-e086-bd64547a093a

+++ non-GPL/d_rcd_sym_v3.h
@@ -45,7 +45,7 @@ class MODEL_BUILT_IN_RCD_SYM_V3 : public MODEL_BUILT_IN_RCD_SYM {
 
     long double __uin_iter(long double& uin,  double cur, double E, const COMMON_COMPONENT* cc)const;
     long double __E(double uin, long double cur, const COMMON_COMPONENT* cc)const;
-    long double __Edu(double uin, long double cur, const COMMON_COMPONENT* cc)const;
+    long double __Edu(long double uin, long double cur, const COMMON_COMPONENT* cc)const;
 };
 /*--------------------------------------------------------------------------*/
 class DEV_BUILT_IN_RCD_SYM_V3 : public DEV_BUILT_IN_RCD{
